feat(registry): integrate service registration into ingestion flow
- Add service_names field to HandleResponse for tracking extracted services
- Implement extract_service_names() to collect unique services from transformed data
- Use ctx.wait_until() for fire-and-forget service registration
- Create WasmRegistrySender in background task to register services
- Fix clippy warning for unnecessary cast in cache.rs

+++ src/registry/cache.rs
@@ -83,7 +83,7 @@ pub fn refresh(services: Vec<String>) {
 /// Uses worker::Date::now() on WASM, SystemTime on native.
 #[cfg(target_arch = "wasm32")]
 fn current_time_ms() -> u64 {
-    worker::Date::now().as_millis() as u64
+    worker::Date::now().as_millis()
 }
 
 #[cfg(not(target_arch = "wasm32"))]
